Task cookie control (#128)
* WIP - This version allows you to add an analytic tool.
I have tested this with Google Analytics and it seems to work locally but having some issues with the edit and remove functions inside the management tool.

Committing to GitHub as stopping for weekend.

* WIP - this commit fixes the edit and remove options.
Still not complete but a more functional demo.

* Removed cookie sections.
Added ability to limit to specific versions for analytics.
Updated database.
WIP end of day commit

* Bootstrap updated to latest version 5.3.0 (#111)

* Bootstrap updated to latest version 5.3.0

* Links to CSS updated

* Added Microsoft Application Insights (client side)
Ready for testing

* Remove migrations to create single migration.

* Consolidate script migrations

* Null object causing an error when trying to create an analytic reference

* Change to get Application Insights script to return. Can't use the @scripts within the context of _AnalyticsPartial.

* Moved the application insights key directly into the map Index.cshtml as it doesn't seem to load if inserted just after the body tag.

Seems to be working now when debugging.

* Added extra options to allow support for Civica Cookie Control which is primarily configured using Google Tag Manager.

* Further tweak to get ICommonRepository to show as no change.

* Update to format and order of cookie preference button.

* Correction requested via GitHub

* Remove old cookie control button reference

* Move function so it fires after DOM load

* Link text colour now specified (#123)

* Updates based on comments on pull request (wording).
Added a global function to allow the CookieControl to be attached to the button post load. This is because it is not always present at DOM load as GTM injects it post load.

Not sure the typescript / webpack method is sound here but we can revise it in a later pull request.

* The map will not load if application insights fails to load.
Added a handler for this situation and fixed it so that it doesn't happen anyway. Handler still needed in case of miss configuration by the user.

* The map will not load if application insights fails to load.
Added a handler for this situation and fixed it so that it doesn't happen anyway. Handler still needed in case of miss configuration by the user.

* Code reviewed to make it more consistent with layer category controller.
We were going to do this as a separate pull request but due to delays updating a setting in the cookie control it is now completed so including in current pull request.

---------

Co-authored-by: Lucy B <114388582+VulpesFerrilata@users.noreply.github.com>